update readme and add some files for future purpose

+++ Spiral/Map.swift
@@ -11,7 +11,7 @@ import SpriteKit
 class Map: SKShapeNode {
     let spacing:CGFloat = 35
     var points:[CGPoint] = []
-    convenience init(origin:CGPoint,layer:CGFloat){
+    convenience init(origin:CGPoint,layer:Int){
         
         var x:CGFloat = origin.x
         var y:CGFloat = origin.y
@@ -19,17 +19,17 @@ class Map: SKShapeNode {
         self.init()
         CGPathMoveToPoint(path, nil, x, y)
         points.append(CGPointMake(x, y))
-        for index in 1..<layer{
-            y-=spacing*(2*index-1)
+        for index in 1..<layer {
+            y-=spacing*(2*CGFloat(index)-1)
             CGPathAddLineToPoint(path, nil , x, y)
             points.append(CGPointMake(x, y))
-            x-=spacing*(2*index-1)
+            x-=spacing*(2*CGFloat(index)-1)
             CGPathAddLineToPoint(path, nil , x, y)
             points.append(CGPointMake(x, y))
-            y+=spacing*2*index
+            y+=spacing*2*CGFloat(index)
             CGPathAddLineToPoint(path, nil , x, y)
             points.append(CGPointMake(x, y))
-            x+=spacing*2*index
+            x+=spacing*2*CGFloat(index)
             CGPathAddLineToPoint(path, nil , x, y)
             points.append(CGPointMake(x, y))
         }
